test: Adds a test to check failure on FIPS
A test has been added to the parallel directory which merely exits with
the specified error message on fips. The parallel.status file also has
been modified to mark this test as flaky on fips so, the default test
suite will work without any modifications. This test also serves as
documentation of the added feature metioned below. Inline documentation
has been added to both the files.

Fixes: https://github.com/nodejs/node/issues/14746